// src/app/model/test2.ts
export interface test2 {
    id3: number;
    inputDate: string;
    inputDay: string;
    inputTime: string;
    inputLocation: string;
    inputCO2: string;
    inputTemperature: string;
    inputHumidity: string;
    inputSO2: string;
    inputNOx: string;
    inputPM: string;
    inputAirVelocity: string;
}